# docs/swagger.yaml
definitions:
...
  api.BlockIPRequest:
    properties:
      alertId:
        description: 'Optional: for audit trail'
        type: string
      ip:
        type: string
      reason:
        description: Optional
        type: string
    type: object
  api.BlockIPResponse:
    properties:
      action:
        type: string
      message:
        type: string
      success:
        type: boolean
      target:
        type: string
    type: object
...
  api.GetRemediationActionsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/api.RemediationActionRecord'
        type: array
      success:
        type: boolean
    type: object
...
  api.HuntIOCsRequest:
    properties:
      alertId:
        description: 'Optional: for context'
        type: string
      iocs:
        items:
          $ref: '#/definitions/api.IOCEntry'
        type: array
    type: object
  api.HuntIOCsResponse:
    properties:
      action:
        type: string
      huntId:
        type: string
      iocCount:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
  api.IOCEntry:
    properties:
      type:
        description: ip, domain, hash, url
        type: string
      value:
        type: string
    type: object
...
  api.RemediationActionRecord:
    properties:
      actionId:
        type: string
      actionType:
        description: block-ip, unblock-ip, hunt-iocs
        type: string
      alertId:
        type: string
      completedAt:
        type: string
      reason:
        type: string
      requestedAt:
        type: string
      requestedBy:
        type: string
      status:
        description: pending, completed, failed
        type: string
      target:
        type: string
    type: object
  api.RemediationError:
    properties:
      code:
        type: string
      error:
        type: string
    type: object
...
  api.UnblockIPRequest:
    properties:
      alertId:
        description: 'Optional: for audit trail'
        type: string
      ip:
        type: string
      reason:
        description: Optional
        type: string
    type: object
  api.UnblockIPResponse:
    properties:
      action:
        type: string
      message:
        type: string
      success:
        type: boolean
      target:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/hunt/iocs:
    post:
      consumes:
      - application/json
      description: |-
        Initiate a threat hunt for specified indicators of compromise.
        This is a placeholder endpoint that logs the hunt request.
        In production, this would trigger searches across SIEM, EDR, and other security tools.
      parameters:
      - description: Hunt IOCs request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.HuntIOCsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: IOC hunt initiated
          schema:
            $ref: '#/definitions/api.HuntIOCsResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/api.RemediationError'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.RemediationError'
      security:
      - BearerAuth: []
      summary: Hunt for IOCs
      tags:
      - remediation
...
  /api/v1/remediation/actions/{alertId}:
    get:
      description: |-
        Retrieve all remediation actions taken for a specific alert.
        This is a placeholder endpoint that returns mock data.
      parameters:
      - description: Alert ID
        in: path
        name: alertId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Remediation actions list
          schema:
            $ref: '#/definitions/api.GetRemediationActionsResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/api.RemediationError'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Alert not found
          schema:
            $ref: '#/definitions/api.RemediationError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.RemediationError'
      security:
      - BearerAuth: []
      summary: Get remediation actions for alert
      tags:
      - remediation
  /api/v1/remediation/block-ip:
    post:
      consumes:
      - application/json
      description: |-
        Block a malicious IP address. This is a placeholder endpoint that logs the remediation request.
        In production, this would integrate with a firewall, EDR, or SOAR platform.
      parameters:
      - description: Block IP request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.BlockIPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: IP blocked successfully
          schema:
            $ref: '#/definitions/api.BlockIPResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/api.RemediationError'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.RemediationError'
      security:
      - BearerAuth: []
      summary: Block IP address
      tags:
      - remediation
  /api/v1/remediation/unblock-ip:
    post:
      consumes:
      - application/json
      description: |-
        Unblock a previously blocked IP address. This is a placeholder endpoint that logs the remediation request.
        In production, this would integrate with a firewall, EDR, or SOAR platform.
      parameters:
      - description: Unblock IP request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.UnblockIPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: IP unblocked successfully
          schema:
            $ref: '#/definitions/api.UnblockIPResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/api.RemediationError'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.RemediationError'
      security:
      - BearerAuth: []
      summary: Unblock IP address
      tags:
      - remediation
...
swagger: "2.0"